+++ resources/views/layouts/app.blade.php
@@ -7,7 +7,7 @@
 
         <x-robots-meta />
 
-        <title>{{ config('app.name', 'Laravel') }}</title>
+        <title>{{ $pageTitle }}</title>
 
         <!-- Scripts -->
         @vite(['resources/css/app.css', 'resources/js/app.js'])