+++ resources/views/welcome.blade.php
@@ -6,7 +6,7 @@
 
         <x-robots-meta />
 
-        <title>{{ config('app.name', 'Laravel') }}</title>
+        <title>{{ config('app.name') }}</title>
 
         <!-- Styles / Scripts -->
         @if (file_exists(public_path('build/manifest.json')) || file_exists(public_path('hot')))